refactor: removendo classes sem uso e movendo classes de pacote
removendo classes sem uso e movendo classes de pacote para não ficarem perdidas na arch, adicionado anotações do swagger para autenticação que estava sem e atualizando a collection do swagger
