+++ src/main/java/br/com/fiap/restaurantusersapi/infrastructure/adapters/inbound/rest/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package br.com.fiap.restaurantusersapi.infrastructure.adapters.inbound.rest;
+package br.com.fiap.restaurantusersapi.infrastructure.adapters.inbound.rest.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
